moved logic for creating array of avaliable movement directions to get to doors from GameLoop to the Room class.
